# spec/openapi.json
{
  "components": {
    "responses": {
      "APIError": {
        "content": {
          "application/json": {
            "schema": {
              "externalDocs": {
                "url": "https://pkg.go.dev/encore.dev/beta/errs#Error"
              },
              "properties": {
                "code": {
                  "description": "Error code",
                  "example": "not_found",
                  "externalDocs": {
                    "url": "https://pkg.go.dev/encore.dev/beta/errs#ErrCode"
                  },
                  "type": "string"
                },
                "details": {
                  "description": "Error details",
                  "type": "object"
                },
                "message": {
                  "description": "Error message",
                  "type": "string"
                }
              },
              "title": "APIError",
              "type": "object"
            }
          }
        },
        "description": "Error response"
      }
    },
...
  },
  "info": {
    "description": "Generated by encore",
    "title": "API for ewix8",
    "version": "1",
    "x-logo": {
      "altText": "Encore logo",
      "backgroundColor": "#EEEEE1",
      "url": "https://encore.dev/assets/branding/logo/logo-black.png"
    }
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/impersonate/start": {
      "post": {
        "operationId": "POST:skyforge.AdminImpersonateStart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "username": {
                    "type": "string"
                  }
                },
                "required": [
                  "username"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "actorUsername": {
                      "type": "string"
                    },
                    "effectiveUsername": {
                      "type": "string"
                    },
                    "impersonating": {
                      "type": "boolean"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "effectiveUsername",
                    "actorUsername",
                    "impersonating"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
...
    "/api/admin/impersonate/stop": {
      "post": {
        "operationId": "POST:skyforge.AdminImpersonateStop",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "effectiveUsername": {
                      "type": "string"
                    },
                    "impersonating": {
                      "type": "boolean"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "effectiveUsername",
                    "impersonating"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
...
    "/api/session": {
      "get": {
        "operationId": "GET:skyforge.Session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "actorUsername": {
                      "type": "string"
                    },
                    "authenticated": {
                      "type": "boolean"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "groups": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "impersonating": {
                      "type": "boolean"
                    },
                    "isAdmin": {
                      "type": "boolean"
                    },
                    "username": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "authenticated",
                    "username",
                    "displayName",
                    "email",
                    "groups",
                    "isAdmin",
                    "actorUsername",
                    "impersonating"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      },
      "head": {
        "operationId": "HEAD:skyforge.SessionHead",
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
    "/api/session/traefik": {
      "get": {
        "operationId": "GET:skyforge.SessionTraefik",
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "summary": "SessionTraefik is a Skyforge SSO gate compatible endpoint.\n"
      },
      "head": {
        "operationId": "HEAD:skyforge.SessionTraefikHead",
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "summary": "SessionTraefikHead is a Skyforge SSO gate compatible endpoint (HEAD).\n"
      }
    },
...
    "/auth/login": {
      "post": {
        "operationId": "POST:skyforge.AuthLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "POST:skyforge.AuthLogout",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
    "/auth/logout-all": {
      "post": {
        "operationId": "POST:skyforge.AuthLogoutAll",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "POST:skyforge.AuthRefresh",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "actorUsername": {
                      "type": "string"
                    },
                    "authenticated": {
                      "type": "boolean"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "groups": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "impersonating": {
                      "type": "boolean"
                    },
                    "isAdmin": {
                      "type": "boolean"
                    },
                    "username": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "authenticated",
                    "username",
                    "displayName",
                    "email",
                    "groups",
                    "isAdmin",
                    "actorUsername",
                    "impersonating"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
    "/auth/validate-token": {
      "post": {
        "operationId": "POST:skyforge.AuthValidateToken",
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "Skyforge (deployed behind Traefik)",
      "url": "/api/skyforge"
    },
    {
      "description": "Encore local dev environment",
      "url": "http://localhost:4000"
    }
  ]
}